% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/yo.R
\name{connpass2yo}
\alias{connpass2yo}
\title{Connpass2Yo}
\usage{
connpass2yo(event_id, to, apikey)
}
\arguments{
\item{event_id}{Connpass event_id}

\item{to}{Username who you want}

\item{apitoken}{Your Yo API token}
}
\description{
\code{yo} yo all subscriber or a specific subscriber.
}
\details{
Before yo(), you need your API key. Please visit http://yoapi.justyo.co/
}
\examples{
\dontrun{
# connpass event
# e.g.) http://zokupata.connpass.com/event/9804/
event2yo(9804, "USERNAME", options()$YO_KEY)
}
}

