% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yo.R
\name{doh}
\alias{doh}
\title{terminate chains of piped commands}
\usage{
doh(x)
}
\arguments{
\item{x}{- an argument}
}
\value{
x - unchanged
}
\description{
doh is useful in terminating long chains of piped commands. It simplifies the debugging
of these long chains by allowing the programmer to comment-out lines of code.
'doh' is pronounced as by the TV-character Homer Simpson.
See also: yo and ruhroh
}
\examples{
library(dplyr)
spray_data <- InsectSprays \%>\% as_data_frame

## simply comment-out lines to remove from chains of piped commands
spray_data \%>\%
  #group_by(spray) \%>\%
  #summarize(mean_count = mean(count)) \%>\%
  doh

spray_data \%>\%
  group_by(spray) \%>\%
  summarize(mean_count = mean(count)) \%>\%
  doh

}
\keyword{doh}

