% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yo.R
\name{ruhroh}
\alias{ruhroh}
\title{useful in terminating long chains of piped commands.
 'ruhroh" is pronounced as by Scooby-Doo}
\usage{
ruhroh(x)
}
\arguments{
\item{x}{- an argument}
}
\value{
x - unchanged
}
\description{
useful in terminating long chains of piped commands.
 'ruhroh" is pronounced as by Scooby-Doo
}
\examples{
library(dplyr)
spray_data <- InsectSprays \%>\% as_data_frame

## simply comment-out lines to remove from chains of piped commands
spray_data \%>\%
  #group_by(spray) \%>\%
  #summarize(mean_count = mean(count)) \%>\%
  ruhroh

spray_data \%>\%
  group_by(spray) \%>\%
  summarize(mean_count = mean(count)) \%>\%
  ruhroh

}
\keyword{ruhroh}

