% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yo.R
\name{yo}
\alias{yo}
\title{yo
A function useful in terminating long chains of piped commands;
 it simplifies debugging by enabling programmer to simply "comment out" lines.
 yo is intended to be used with the dplyr package, but does not depend on it.
 yo is pronounced as by Aaron Paul's character, Jesse Pinkman, in "Breaking Bad"}
\usage{
yo(x)
}
\arguments{
\item{x}{- an argument}
}
\value{
x - unchanged
}
\description{
yo
A function useful in terminating long chains of piped commands;
 it simplifies debugging by enabling programmer to simply "comment out" lines.
 yo is intended to be used with the dplyr package, but does not depend on it.
 yo is pronounced as by Aaron Paul's character, Jesse Pinkman, in "Breaking Bad"
}
\examples{
library(dplyr)
spray_data <- InsectSprays \%>\% as_data_frame

## simply comment-out lines to remove from chains of piped commands
spray_data \%>\%
  #group_by(spray) \%>\%
  #summarize(mean_count = mean(count)) \%>\%
  yo

spray_data \%>\%
  group_by(spray) \%>\%
  #summarize(mean_count = mean(count)) \%>\%
  yo

spray_data \%>\%
  group_by(spray) \%>\%
  summarize(mean_count = mean(count)) \%>\%
  yo

}
\keyword{yo}

