% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yo.R
\name{doh}
\alias{doh}
\title{useful in terminating long chains of piped commands.
 'doh' is pronounced as by Homer Simpson}
\usage{
doh(x)
}
\arguments{
\item{x}{- an argument}
}
\value{
x - unchanged
}
\description{
useful in terminating long chains of piped commands.
 'doh' is pronounced as by Homer Simpson
}
\examples{
library(dplyr)
spray_data <- InsectSprays \%>\% as_data_frame

## simply comment-out lines to remove from chains of piped commands
spray_data \%>\%
  #group_by(spray) \%>\%
  #summarize(mean_count = mean(count)) \%>\%
  doh

spray_data \%>\%
  group_by(spray) \%>\%
  summarize(mean_count = mean(count)) \%>\%
  doh

}
\keyword{doh}

