% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yo.R
\name{ruhroh}
\alias{ruhroh}
\title{terminate chains of piped commands}
\usage{
ruhroh(x)
}
\arguments{
\item{x}{- an argument}
}
\value{
x - unchanged
}
\description{
ruhroh is useful in terminating long chains of piped commands. It simplifies the debugging
of these long chains by allowing the programmer to comment-out lines of code.
'ruhroh' is pronounced as by the TV-character Scooby-Doo.
See also: yo, ruhroh
}
\examples{
library(dplyr)
spray_data <- InsectSprays \%>\% as_data_frame

## simply comment-out lines to remove from chains of piped commands
spray_data \%>\%
  #group_by(spray) \%>\%
  #summarize(mean_count = mean(count)) \%>\%
  ruhroh

spray_data \%>\%
  group_by(spray) \%>\%
  summarize(mean_count = mean(count)) \%>\%
  ruhroh

}
\keyword{ruhroh}

