% Generated by roxygen2 (4.0.1): do not edit by hand
\name{yo}
\alias{yo}
\title{Just Yo}
\usage{
yo(apitoken = NULL, username = NULL)
}
\arguments{
\item{apitoken}{Your Yo API token}

\item{username}{Username who you want}
}
\value{
status code and status message
}
\description{
\code{yo} yo all subscriber or a specific subscriber.
}
\details{
Before yo(), you need your API key. Please visit http://yoapi.justyo.co/
}
\examples{
\dontrun{
# Yo all subscriber.
beep("Your API token")
# Yo a specific subscriber
beep("Your API token", "username")
}
}

